requirements-tox.txt
# Common requirements for envs created by tox
# Additional, version-dependent requirements can be found in tox.ini
coverage==4.5.4
mock==3.0.5
pytest-cov==2.8.1
pytest-pythonpath==0.7.3
